+++ Controls/_scroll.less
@@ -2,5 +2,5 @@
 @import "Controls/_scroll/Scroll/_Scroll";
 @import "Controls/_scroll/ContainerBase/_ContainerBase";
 @import "Controls/_scroll/Container/_Container";
-@import "Controls/_scroll/StickyHeader/_StickyHeader/_StickyHeader";
+@import "Controls/_scroll/StickyHeader/_StickyHeader";
 @import "Controls/_scroll/IntersectionObserver/EdgeIntersectionContainer";
