+++ Controls/_scroll/StickyHeader/Group.ts
@@ -11,7 +11,7 @@ import {
     getGapFixSize
 } from 'Controls/_scroll/StickyHeader/Utils';
 import template = require('wml!Controls/_scroll/StickyHeader/Group');
-import {SHADOW_VISIBILITY} from './_StickyHeader';
+import {SHADOW_VISIBILITY} from '../StickyHeader';
 import {RegisterUtil, UnregisterUtil} from 'Controls/event';
 import fastUpdate from './FastUpdate';
 
@@ -19,7 +19,7 @@ import fastUpdate from './FastUpdate';
  * Allows you to combine sticky headers with the same behavior. It is necessary if you need to make
  * several headers fixed at the same level, which should simultaneously stick and stick out.
  * Behaves like one fixed header.
- * 
+ *
  * @remark
  * Полезные ссылки:
  * * <a href="https://github.com/saby/wasaby-controls/blob/rc-20.4000/Controls-default-theme/aliases/_scroll.less">переменные тем оформления</a>
