+++ Controls/scroll.ts
@@ -31,7 +31,7 @@
 import Container = require('Controls/_scroll/Container');
 import ContainerNew from 'Controls/_scroll/ContainerNew';
 import Watcher = require('Controls/_scroll/Scroll/Watcher');
-import StickyHeader = require('Controls/_scroll/StickyHeader');
+export {default as StickyHeader} from 'Controls/_scroll/StickyHeader';
 import _Scrollbar = require('Controls/_scroll/Scroll/Scrollbar');
 import _scrollContext = require('Controls/_scroll/Scroll/Context');
 import _stickyHeaderContext = require('Controls/_scroll/StickyHeader/Context');
@@ -53,7 +53,6 @@ export {
    Container,
    ContainerNew as _ContainerNew,
    Watcher,
-   StickyHeader,
    _Scrollbar,
    _scrollContext,
    _stickyHeaderContext,
