+++ tests/ControlsUnit/StickyHeader/Template.test.js
@@ -4,7 +4,7 @@ define(
       'Env/Env',
       'ControlsUnit/resources/TemplateUtil',
       'Controls/_scroll/StickyHeader/Utils',
-      'Controls/_scroll/StickyHeader/_StickyHeader'
+      'Controls/_scroll/StickyHeader'
    ],
    function(scroll, Env, TemplateUtil, StickyHeaderUtils, _StickyHeaderLib) {
 
@@ -57,21 +57,6 @@ define(
 
                assert.equal(template(inst), '<div class="controls-background-default_theme-default"><div>testing the template</div></div>');
             });
-
-            it('The browser does support sticky', function() {
-               inst._isStickySupport = true;
-               inst._options.content = TemplateUtil.content;
-               sinon.stub(_StickyHeader.prototype, '_getComputedStyle').returns({ });
-
-               template(inst, function(result) {
-                  assert.equal(result, '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader__background controls-StickyHeader_position">' +
-                     '<div class="controls-StickyHeader__observationTargetTop" style="top: -3px;"></div>' +
-                     '<div class="controls-StickyHeader__content">testing the template</div>' +
-                     '<div class="controls-StickyHeader__observationTargetBottom" style="bottom: -3px;"></div>' +
-                     '</div>');
-               });
-               sinon.restore();
-            });
          });
 
          describe('_StickyHeader', function() {
@@ -89,7 +74,9 @@ define(
                inst._reverseOffsetStyle = ctrl._reverseOffsetStyle;
                inst._getBottomShadowStyle = ctrl._getBottomShadowStyle;
                inst._getNormalizedContainer = ctrl._getNormalizedContainer;
-               inst._getComputedStyle = () => { return {}; };
+               inst._getComputedStyle = () => {
+                  return {};
+               };
                inst._scrollState = {};
                template = TemplateUtil.clearTemplate(ctrl._template);
             });
@@ -98,11 +85,14 @@ define(
                inst._isMobilePlatform = false;
                inst._model.fixedPosition = 'top';
                inst._options.theme = 'default';
-               inst._options.content = function() {return ''};
+               inst._options.content = function() {
+                  return ''
+               };
+               inst._isStickySupport = true;
 
                inst._updateStyles();
 
-               assert.equal(template(inst),  '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="top: 0px;z-index: 2;">' +
+               assert.equal(template(inst), '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="top: 0px;z-index: 2;">' +
                   '<div class="controls-StickyHeader__observationTargetTop" style="top: -3px;"></div>' +
                   '<div class="controls-StickyHeader__observationTargetBottom" style="bottom: -3px;"></div>' +
                   '</div>');
@@ -114,15 +104,18 @@ define(
                sandbox.stub(StickyHeaderUtils, 'getGapFixSize').returns(1);
                inst._model.fixedPosition = 'top';
                sandbox.replace(inst, '_getComputedStyle', function() {
-                  return { paddingTop: '0px' };
+                  return {paddingTop: '0px'};
                });
-               inst._container = { style: { paddingTop: '' } };
+               inst._container = {style: {paddingTop: ''}};
                inst._options.theme = 'default';
-               inst._options.content = function() {return ''};
+               inst._options.content = function() {
+                  return ''
+               };
+               inst._isStickySupport = true;
 
                inst._updateStyles();
 
-               assert.equal(template(inst),  '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="top: -1px;padding-top:1px;margin-top: -1px;z-index: 2;">' +
+               assert.equal(template(inst), '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="top: -1px;padding-top:1px;margin-top: -1px;z-index: 2;">' +
                   '<div class="controls-StickyHeader__observationTargetTop" style="top: -3px;"></div>' +
                   '<div class="controls-StickyHeader__observationTargetBottom" style="bottom: -3px;"></div>' +
                   '</div>');
@@ -131,11 +124,14 @@ define(
 
             it('Move the header', function() {
                inst._options.theme = 'default';
-               inst._options.content = function() {return ''};
+               inst._options.content = function() {
+                  return ''
+               };
+               inst._isStickySupport = true;
 
                inst._updateStyles();
 
-               assert.equal(template(inst),  '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="top: 0px;">' +
+               assert.equal(template(inst), '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="top: 0px;">' +
                   '<div class="controls-StickyHeader__observationTargetTop" style="top: -3px;"></div>' +
                   '<div class="controls-StickyHeader__observationTargetBottom" style="bottom: -3px;"></div>' +
                   '</div>');
@@ -144,11 +140,14 @@ define(
             it('Move the bottom', function() {
                inst._options.position = 'bottom';
                inst._options.theme = 'default';
-               inst._options.content = function() {return ''};
+               inst._options.content = function() {
+                  return ''
+               };
+               inst._isStickySupport = true;
 
                inst._updateStyles();
 
-               assert.equal(template(inst),  '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="bottom: 0px;">' +
+               assert.equal(template(inst), '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="bottom: 0px;">' +
                   '<div class="controls-StickyHeader__observationTargetTop" style="top: -3px;"></div>' +
                   '<div class="controls-StickyHeader__observationTargetBottom" style="bottom: -3px;"></div>' +
                   '</div>');
@@ -157,10 +156,11 @@ define(
             it('Added content', function() {
                inst._options.content = TemplateUtil.content;
                inst._options.theme = 'default';
+               inst._isStickySupport = true;
 
                inst._updateStyles();
 
-               assert.equal(template(inst),   '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="top: 0px;">' +
+               assert.equal(template(inst), '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="top: 0px;">' +
                   '<div class="controls-StickyHeader__observationTargetTop" style="top: -3px;"></div>' +
                   '<div class="controls-StickyHeader__content">testing the template</div>' +
                   '<div class="controls-StickyHeader__observationTargetBottom" style="bottom: -3px;"></div>' +
@@ -174,20 +174,22 @@ define(
                inst._options.fixedZIndex = 1;
                inst._options.content = TemplateUtil.content;
                inst._options.theme = 'default';
+               inst._isStickySupport = true;
 
                inst._updateStyles();
 
-               assert.equal(template(inst),  '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="top: 0px;z-index: 1;">' +
+               assert.equal(template(inst), '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="top: 0px;z-index: 1;">' +
                   '<div class="controls-StickyHeader__observationTargetTop" style="top: -3px;"></div>' +
                   '<div class="controls-StickyHeader__content">testing the template</div>' +
                   '<div class="controls-StickyHeader__observationTargetBottom" style="bottom: -3px;"></div>' +
-                   '<div class="controls-Scroll__shadow controls-StickyHeader__shadow-bottom controls-Scroll__shadow_horizontal"></div>' +
-                   '</div>');
+                  '<div class="controls-Scroll__shadow controls-StickyHeader__shadow-bottom controls-Scroll__shadow_horizontal"></div>' +
+                  '</div>');
             });
 
             it('The header is fixed, the shadow should be', function() {
                inst._scrollState.verticalPosition = 'start';
                inst._isFixed = true;
+               inst._isStickySupport = true;
                inst._model.fixedPosition = 'bottom';
                inst._options.fixedZIndex = 2;
                inst._options.position = 'bottom';
@@ -196,7 +198,7 @@ define(
 
                inst._updateStyles();
 
-               assert.equal(template(inst),  '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="bottom: 0px;z-index: 2;">' +
+               assert.equal(template(inst), '<div class="controls-StickyHeader controls-background-default_theme-default controls-StickyHeader_position" style="bottom: 0px;z-index: 2;">' +
                   '<div class="controls-Scroll__shadow controls-StickyHeader__shadow-top controls-Scroll__shadow_horizontal"></div>' +
                   '<div class="controls-StickyHeader__observationTargetTop" style="top: -3px;"></div>' +
                   '<div class="controls-StickyHeader__content">testing the template</div>' +
